move forecast classes to separate modules
